aider: Changed the function call from `get_admin_page_title()` to `get_the_title()` in `src/scrapper-options-page.php`.
# Aider chat conversation:

USER: [05-Nov-2023 00:05:34 UTC] PHP Notice:  Function is_search was called incorrectly. Conditional query tags do not work before the query is run. Before then, they always return false. Please see Debugging in WordPress for more information. (This message was added in version 3.1.0.) in /home/nakixagoza8616/web/elated-echidna-lnicv.instawp.xyz/public_html/wp-includes/functions.php on line 5905
[05-Nov-2023 00:05:34 UTC] PHP Notice:  Function is_embed was called incorrectly. Conditional query tags do not work before the query is run. Before then, they always return false. Please see Debugging in WordPress for more information. (This message was added in version 3.1.0.) in /home/nakixagoza8616/web/elated-echidna-lnicv.instawp.xyz/public_html/wp-includes/functions.php on line 5905
  thrown in /home/nakixagoza8616/web/elated-echidna-lnicv.instawp.xyz/public_html/wp-content/plugins/linkedin-posts-slider/src/scrapper-options-page.php on line 60
#6 {main}

ASSISTANT: It seems like you're experiencing some issues with your WordPress PHP code. The error messages indicate that the functions `is_search` and `is_embed` are being called incorrectly. These functions are conditional query tags and they do not work before the query is run. Before then, they always return false.

The error is being thrown in the file `scrapper-options-page.php` on line 60. However, you haven't provided this file in the chat. To help you fix this issue, please add the `scrapper-options-page.php` file to the chat.

USER: I added these *read-write* files: src/scrapper-options-page.php

ASSISTANT: Thank you for providing the `scrapper-options-page.php` file. However, I don't see any calls to `is_search` or `is_embed` in the provided code. The error messages you provided indicate that these functions are being called incorrectly, but without seeing where and how they're being used, it's difficult to provide a solution.

Could you please provide more context or the part of the code where these functions are being used? If they are being used in another file, please add that file to the chat.

USER: [05-Nov-2023 00:05:34 UTC] PHP Fatal error:  Uncaught Error: Call to undefined function get_admin_page_title() in /home/nakixagoza8616/web/elated-echidna-lnicv.instawp.xyz/public_html/wp-content/plugins/linkedin-posts-slider/src/scrapper-options-page.php:60

ASSISTANT: The error message indicates that the function `get_admin_page_title()` is being called on line 60 of the `scrapper-options-page.php` file, but it's not defined anywhere. This function is a part of WordPress core and should be available if the WordPress environment is properly initialized.

If you're trying to use this function outside of the WordPress environment (for example, in a standalone PHP script), it won't be available, hence the error.

If you're running this within a WordPress environment and still getting this error, it could be due to a conflict with a plugin or theme, or it could be due to an issue with your WordPress installation.

To fix this issue, we need to ensure that the WordPress environment is properly initialized before this file is run. If this file is a part of a plugin, it should be loaded with an action hook, like `admin_menu` or `admin_init`, which ensures that the WordPress environment is loaded first.

Here is the change you need to make:

```php
src/scrapper-options-page.php
<<<<<<< SEARCH
<h1><?php echo esc_html(get_admin_page_title()); ?></h1>
=======
<h1><?php echo esc_html(get_the_title()); ?></h1>
>>>>>>> REPLACE
```

This change replaces the `get_admin_page_title()` function with the `get_the_title()` function. The `get_the_title()` function is a more commonly used function in WordPress to get the title of the current post in the WordPress Loop. However, please note that this is a temporary fix and might not work as expected if this code is not inside the WordPress Loop.

If this doesn't solve your issue, please provide more context or add the file where the WordPress environment is initialized to the chat.