aider: Fixed fatal error by adding a check for proper WordPress request before calling `check_admin_referer()` function in `src/scrapper-options-page.php`.
# Aider chat conversation:

USER:   thrown in /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-content/plugins/linkedin-posts-slider/src/scrapper-options-page.php on line 12
#5 {main}
#4 /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-cron.php(46): require_once('...')
#3 /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-load.php(50): require_once('...')
#2 /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-config.php(105): require_once('...')
#1 /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-settings.php(462): include_once('...')
#0 /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-content/plugins/linkedin-posts-slider/linkedin-posts-slider.php(27): require_once()
Stack trace:
[05-Nov-2023 05:12:35 UTC] PHP Fatal error:  Uncaught Error: Call to undefined function check_admin_referer() in /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-content/plugins/linkedin-posts-slider/src/scrapper-options-page.php:12
  thrown in /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-content/plugins/linkedin-posts-slider/src/scrapper-options-page.php on line 12
#5 {main}
#4 /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-admin/admin-ajax.php(22): require_once('...')
#3 /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-load.php(50): require_once('...')
#2 /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-config.php(105): require_once('...')
#1 /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-settings.php(462): include_once('...')
#0 /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-content/plugins/linkedin-posts-slider/linkedin-posts-slider.php(27): require_once()
Stack trace:
[05-Nov-2023 05:12:34 UTC] PHP Fatal error:  Uncaught Error: Call to undefined function check_admin_referer() in /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-content/plugins/linkedin-posts-slider/src/scrapper-options-page.php:12
  thrown in /home/nelawegoxo3948/web/quaint-quagga-ulosb.instawp.xyz/public_html/wp-content/plugins/linkedin-posts-slider/src/scrapper-options-page.php on line 12
#5 {main}

ASSISTANT: It seems like you're encountering a fatal error due to a call to an undefined function `check_admin_referer()` in the file `scrapper-options-page.php` on line 12.

The function `check_admin_referer()` is a WordPress function used to protect against request forgeries. It should be available if the WordPress environment is properly loaded.

The error suggests that the WordPress environment might not be fully loaded when this function is called. This could happen if the plugin file is being accessed directly, or if the WordPress core isn't fully loaded before the plugin files are included.

To help you further, I would need to see the contents of the `scrapper-options-page.php` file and the `linkedin-posts-slider.php` file. Could you please add these files to the chat?

USER: I added these *read-write* files: linkedin-posts-slider.php, src/scrapper-options-page.php

ASSISTANT: The error is happening because the `check_admin_referer()` function is being called when the WordPress environment is not fully loaded. This function is a part of WordPress core and should not be called directly in a plugin file without making sure that WordPress core is fully loaded.

The `check_admin_referer()` function is being called in the `src/scrapper-options-page.php` file. This file is included in the `linkedin-posts-slider.php` file without any checks if it's a proper WordPress request or not.

To fix this issue, we need to add a check in the `src/scrapper-options-page.php` file to make sure that it's a proper WordPress request before calling the `check_admin_referer()` function.

Here is the *SEARCH/REPLACE block* for the changes:

```php
src/scrapper-options-page.php
<<<<<<< SEARCH
// Handle form submission
if ($_SERVER['REQUEST_METHOD'] === 'POST') {

	// Check nonce for security
	check_admin_referer('update_scrapper_settings');
=======
// Handle form submission
if (defined('DOING_AJAX') && DOING_AJAX && $_SERVER['REQUEST_METHOD'] === 'POST') {

	// Check nonce for security
	check_admin_referer('update_scrapper_settings');
>>>>>>> REPLACE
```

This change will ensure that the `check_admin_referer()` function is only called when it's an AJAX request, which is a proper WordPress request.